Allow authentication caller to provide raw hashed challenge
Some uses of FIDO2 authentication do not use a hashed JSON challenge,
but instead their own data blob that is SHA256 hashed in the same
fashion. An example of this is the use of FIDO2 keys with SSH. Provide a
mechanism for indicating that the caller has already calculated the
hash.

Tested with a Go SSH/FIDO2 agent implementation.